backend/requirements.txt
fastapi
uvicorn
langchain
pinecone-client
sentence-transformers
